requirements.txt
Django==4.1.9
djangorestframework==3.14.0
